Make MacOS happy about printf type for tv.tv_usec

+++ src/binance_logger.cpp
@@ -36,7 +36,7 @@ void binance::Logger::write_log( const char *fmt, ... )
 
 
 	sprintf( new_fmt, "%04d-%02d-%02d %02d:%02d:%02d %06ld :%s\n",
-		now->tm_year + 1900, now->tm_mon + 1, now->tm_mday, now->tm_hour, now->tm_min, now->tm_sec, tv.tv_usec, fmt );
+		now->tm_year + 1900, now->tm_mon + 1, now->tm_mday, now->tm_hour, now->tm_min, now->tm_sec, (long)tv.tv_usec, fmt );
 
 	va_start (arg, fmt);
